djlibrary/requirements.txt
2captcha-python==1.1.2
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
distlib==0.3.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
docutils==0.19
drf-yasg==1.21.3
et-xmlfile==1.1.0
filelock==3.7.1
frozenlist==1.3.1
h11==0.13.0
idna==3.3
Markdown==3.4.1
multidict==6.0.2
openpyxl==3.0.10
outcome==1.2.0
Pillow==9.2.0
platformdirs==2.5.2
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
Pyrogram==2.0.35
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
rsa==4.9
selenium==4.4.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
Telethon==1.24.0
TgCrypto==1.2.3
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
virtualenv==20.16.3
wsproto==1.1.0
yarl==1.8.1
